pyproject.toml
[project]
name = "velgarien-platform"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "supabase>=2.0.0",
    "httpx>=0.27.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "Pillow>=10.0.1",
    "cryptography>=41.0.7",
    "python-jose[cryptography]>=3.3.0",
    "structlog>=23.2.0",
    "tenacity>=9.1.2",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "httpx",
    "ruff>=0.3.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T20", "ICN"]

[tool.ruff.lint.per-file-ignores]
"backend/tests/**/*.py" = ["S101", "S105", "S106"]
"backend/routers/**/*.py" = ["B008"]
"backend/dependencies.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
asyncio_mode = "auto"
